Move PostgreSQL store instance to its own file
Remove defunct ServerState data type and parameterise the servant server
Move authHeader prism to Types from Store
Newtype wrapped PostgreSQL connection pool for instance
