+++ oauth2-server.cabal
@@ -24,6 +24,8 @@ Library
                      , Network.OAuth2.Server.Configuration
                      , Network.OAuth2.Server.Statistics
                      , Network.OAuth2.Server.Store
+                     , Network.OAuth2.Server.Store.Base
+                     , Network.OAuth2.Server.Store.PostgreSQL
                      , Network.OAuth2.Server.Types
                      , Network.OAuth2.Server.Types.Common
                      , Network.OAuth2.Server.Types.Scope
